% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe-index.R
\name{hall_of_fame}
\alias{hall_of_fame}
\title{Get Programming Language Hall of Fame Table}
\usage{
hall_of_fame()
}
\value{
A dataframe with Programming Language Hall of Fame Table
}
\description{
Get Programming Language Hall of Fame Table
}
\examples{
hall_of_fame()
}
