% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe-index.R
\name{long_term_history}
\alias{long_term_history}
\title{Get Long Term History table of popular programming languages}
\usage{
long_term_history()
}
\value{
A dataframe with Popular Programming Languages over a long Term History
}
\description{
Get Long Term History table of popular programming languages
}
\examples{
long_term_history()
}
