% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobeindexr-package.R
\docType{package}
\name{tiobeindexr}
\alias{tiobeindexr}
\alias{tiobeindexr-package}
\title{Extract 'TIOBE' Programming Index Tables for the latest month}
\description{
Get latest TIOBE Index Tables. The TIOBE Programming Community index
is an indicator of the popularity of programming languages
from \url{https://www.tiobe.com/tiobe-index/}
}
\section{See Also}{

Useful links:
\itemize{
\item \url{https://www.tiobe.com/tiobe-index/}
\item \url{https://github.com/amrrs/tiobeindexr}
\item Report bugs at \href{https://github.com/amrrs/tiobieindexr/issues}{https://github.com/amrrs/tiobeindexr/issues}
}
}

\author{
AbdulMajedRaja RS(amrrs.data@gmail.com)
}
