% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobieindexr-package.R
\docType{package}
\name{tiobieindexr}
\alias{tiobieindexr}
\alias{tiobieindexr-package}
\title{Extract 'TIOBE' Programming Index Tables for the latest month}
\description{
Get latest TIOBE Index Tables. The TIOBE Programming Community index
is an indicator of the popularity of programming languages
from \url{https://www.tiobe.com/tiobe-index/}
}
\author{
AbdulMajedRaja RS(amrrs.data@gmail.com)
}
