% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe-index.R
\name{top_20}
\alias{top_20}
\title{Get Top 20 values of TIOBIE Index table for the latest month}
\usage{
top_20()
}
\value{
A dataframe with Top 20 Programming languages by their TIOBE Index
        along with a comparison between Last month's and
        this month's ranking and Change in the Ratings
}
\description{
Get Top 20 values of TIOBIE Index table for the latest month
}
\examples{
top_20()
}
