% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiobe-index.R
\name{top_50}
\alias{top_50}
\title{Get Top 50 values of TIOBIE Index table for the latest month}
\usage{
top_50()
}
\value{
A dataframe with Top 50 Programming languages by their TIOBE Index
}
\description{
Get Top 50 values of TIOBIE Index table for the latest month
}
\examples{
top_50()
}
